+++ src/Behaviours/MakeTag.php
@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace A17\EdgeFlush\Behaviours;
 
@@ -7,16 +7,25 @@ use A17\EdgeFlush\Services\Invalidation;
 
 trait MakeTag
 {
-    public function makeModelName(Model $model, string $key = null, array $allowedKeys = []): string|null
-    {
+    public function makeModelName(
+        Model $model,
+        string $key = null,
+        array $allowedKeys = []
+    ): string|null {
         if ($key === 'title') {
-
-
-        info(['------ makeModelName ----', get_class($model), $key, $allowedKeys, $this->keyIsAllowed($key, $allowedKeys), $model->getCDNCacheTag($key)]);
+            info([
+                '------ makeModelName ----',
+                get_class($model),
+                $key,
+                $allowedKeys,
+                $this->keyIsAllowed($key, $allowedKeys),
+                $model->getCDNCacheTag($key),
+            ]);
         }
-        
+
         try {
-            return method_exists($model, 'getCDNCacheTag') && $this->keyIsAllowed($key, $allowedKeys)
+            return method_exists($model, 'getCDNCacheTag') &&
+                $this->keyIsAllowed($key, $allowedKeys)
                 ? $model->getCDNCacheTag($key)
                 : null;
         } catch (\Exception $exception) {