+++ src/EdgeFlush.php
@@ -1,23 +1,32 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace A17\EdgeFlush;
 
+use Illuminate\Http\Request;
+use A17\EdgeFlush\Services\Tags;
+use A17\EdgeFlush\Services\Warmer;
 use Illuminate\Support\Facades\Facade;
+use A17\EdgeFlush\Contracts\CDNService;
+use A17\EdgeFlush\Services\CacheControl;
+use Symfony\Component\HttpFoundation\Response;
 
 /**
- * @method static \Symfony\Component\HttpFoundation\Response makeResponse(Response $response)
- * @method static \A17\EdgeFlush\Contracts\CDNService cdn()
- * @method static \A17\EdgeFlush\Services\CacheControl cacheControl()
- * @method static \A17\EdgeFlush\Services\Tags tags()
- * @method static \A17\EdgeFlush\Services\Warmer warmer()
- * @method static \Illuminate\Http\Request\Request getRequest()
+ * @method static Response makeResponse(Response $response)
+ * @method static CDNService cdn()
+ * @method static CacheControl cacheControl()
+ * @method static Tags tags()
+ * @method static Warmer warmer()
+ * @method static Request getRequest()
  * @method static bool enabled()
  * @method static self instance()
  * @method static self setRequest(Request $request)
+ * @method static bool invalidationServiceIsEnabled()
+ * @method static bool storeTagsServiceIsEnabled()
+ * @method static bool warmerServiceIsEnabled()
  **/
 class EdgeFlush extends Facade
 {
-    protected static function getFacadeAccessor()
+    protected static function getFacadeAccessor(): string
     {
         return 'a17.edge-flush.service';
     }