+++ src/Support/Helpers.php
@@ -1,9 +1,10 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace A17\EdgeFlush\Support;
 
 use Throwable;
-use Illuminate\Support\Str;
+use A17\EdgeFlush\Models\Url;
+use A17\EdgeFlush\Models\Tag;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\Log;
 
@@ -14,19 +15,21 @@ class Helpers
      */
     public static function sanitizeUrl(string $url): string
     {
-        if (config('edge-flush.urls.query.fully_cachable')) {
+        if (Helpers::configBool('edge-flush.urls.query.fully_cachable')) {
             return $url;
         }
 
         $parsed = static::parseUrl($url);
 
-        parse_str($parsed['query'] ?? null, $query);
+        $query = [];
+
+        parse_str($parsed['query'] ?? '', $query);
 
         if (blank($query)) {
             return $url;
         }
 
-        $routes = config('edge-flush.urls.query.allow_routes');
+        $routes = (array) config('edge-flush.urls.query.allow_routes');
 
         $list = $routes[$parsed['path']] ?? null;
 
@@ -34,7 +37,7 @@ class Helpers
             try {
                 $routeName = app('router')
                     ->getRoutes()
-                    ->match(app('request')->create($url))
+                    ->match(request()->create($url))
                     ->getName();
             } catch (\Throwable) {
                 $routeName = '';
@@ -56,7 +59,7 @@ class Helpers
      * Deconstruct and rebuild the URL dropping query parameters
      */
     public static function rewriteUrl(
-        mixed $parameters,
+        array $parameters,
         array|Collection $dropQueries = [],
         string|null $url = null
     ): string {
@@ -64,15 +67,9 @@ class Helpers
 
         $url = static::parseUrl($url);
 
-        if (is_string($parameters)) {
-            $parameters = explode('=', $parameters);
-
-            $parameters = [$parameters[0] => $parameters[1]];
-        }
-
         $query = [];
 
-        parse_str($url['query'] ?? '', $query);
+        parse_str($url['query'] ?? null, $query);
 
         foreach ($parameters as $key => $parameter) {
             $query[$key] = $parameter;
@@ -95,8 +92,8 @@ class Helpers
         $url = $components['scheme'] . '://';
 
         if (
-            !empty($components['username']) &&
-            !empty($components['password'])
+            filled($components['username']) &&
+            filled($components['password'])
         ) {
             $url .=
                 $components['username'] . ':' . $components['password'] . '@';
@@ -105,7 +102,7 @@ class Helpers
         $url .= $components['host'];
 
         if (
-            !empty($components['port']) &&
+            filled($components['port']) &&
             (($components['scheme'] === 'http' && $components['port'] !== 80) ||
                 ($components['scheme'] === 'https' &&
                     $components['port'] !== 443))
@@ -113,34 +110,32 @@ class Helpers
             $url .= ':' . $components['port'];
         }
 
-        if (!empty($components['path'])) {
+        if (filled($components['path'])) {
             $url .= $components['path'];
         }
 
-        if (!empty($components['fragment'])) {
+        if (filled($components['fragment'])) {
             $url .= '#' . $components['fragment'];
         }
 
-        if (!empty($components['query'])) {
+        if (filled($components['query'])) {
             $url .= '?' . http_build_query($components['query']);
         }
 
         return $url;
     }
 
-    public static function parseUrl(mixed $url): array
+    public static function parseUrl(object|array|string|null $url): array
     {
         if (is_array($url)) {
             $url = '';
         }
 
-        if (is_object($url)) {
-            try {
-                /** @throws Throwable */
-                $url = (string) $url;
-            } catch (Throwable) {
-                $url = '';
-            }
+        try {
+            /** @throws Throwable */
+            $url = (string) $url;
+        } catch (Throwable) {
+            $url = '';
         }
 
         /** Check if the string only a domain name **/
@@ -167,14 +162,14 @@ class Helpers
 
     public static function debug(...$data): bool
     {
-        $debugIsOn = config('edge-flush.debug');
+        $debugIsOn = (bool) config('edge-flush.debug', false);
 
         if (!$debugIsOn) {
             return false;
         }
 
         if (blank($data)) {
-            return $debugIsOn;
+            return true;
         }
 
         if (!is_string($data)) {
@@ -185,6 +180,109 @@ class Helpers
 
         Log::debug('[EDGE-FLUSH] ' . $data);
 
-        return $debugIsOn;
+        return true;
+    }
+
+    public static function toString(mixed $string): string
+    {
+        if (is_string($string) || is_numeric($string)) {
+            return (string) $string;
+        }
+
+        if (is_array($string)) {
+            return (string) json_encode($string);
+        }
+
+        if (is_object($string)) {
+            return (string) json_encode($string);
+        }
+
+        return '';
+    }
+
+    public static function toInt(mixed $string): int
+    {
+        if (is_string($string) || is_numeric($string)) {
+            return (int) $string;
+        }
+
+        return 0;
+    }
+
+    public static function toArray(mixed $array): array
+    {
+        if (is_array($array)) {
+            return $array;
+        }
+
+        if ($array instanceof Collection) {
+            return $array->toArray();
+        }
+
+        if (is_string($array) || is_numeric($array)) {
+            return (array) $array;
+        }
+
+        if (is_object($array)) {
+            return (array) $array;
+        }
+
+        return [];
+    }
+
+    public static function toBool(mixed $value): bool
+    {
+        return (bool) $value;
+    }
+
+    public static function configBool(string $key, mixed $default = null): bool
+    {
+        return static::toBool(config($key, $default));
+    }
+
+    public static function configArray(
+        string $key,
+        mixed $default = null
+    ): array|null {
+        if (is_null($value = config($key, $default))) {
+            return null;
+        }
+
+        return static::toArray($value);
+    }
+
+    public static function configString(
+        string $key,
+        mixed $default = null
+    ): string|null {
+        if (is_null($value = config($key, $default))) {
+            return null;
+        }
+
+        return static::toString($value);
+    }
+
+    public static function configInt(
+        string $key,
+        mixed $default = null
+    ): int|null {
+        if (is_null($value = config($key, $default))) {
+            return null;
+        }
+
+        return static::toInt($value);
+    }
+
+    public static function getUrl(mixed $item): string|null
+    {
+        if ($item instanceof Url) {
+            $url = $item->url;
+        } elseif ($item instanceof Tag) {
+            $url = $item->url->url;
+        } else {
+            $url = null;
+        }
+
+        return $url;
     }
 }